+++ js/map.js
@@ -1,5 +1,4 @@
 import {enablePage, enableFilters} from '../js/form.js';
-import {defaultRoomSelector} from '../js/form-validaty.js';
 import {generateCardElement} from '../js/make-card.js'; //функция генерации карточек
 const inputAddress = document.querySelector('#address');
 const resetButton = document.querySelector('.reset__map');
@@ -8,43 +7,44 @@ const lngCenter = (139.7539934567).toFixed(5);
 let lat = +latCenter;
 let lng = +lngCenter;
 
-const showMap = function(loadedCardData) {
-  inputAddress.value = `${lat}, ${lng}`;
-  // /* global L:readonly */
+inputAddress.value = `${lat}, ${lng}`;
+// /* global L:readonly */
 
-  const map = L.map('map-canvas')
-    .on('load', () => {
-      enablePage();
-    })
-    .setView({lat, lng}, 12);
+const map = L.map('map-canvas')
+  .setView({lat, lng}, 12);
 
-  L.tileLayer(
-    'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-    {attribution:
+L.tileLayer(
+  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+  {attribution:
     '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'})
-    .addTo(map);
+  .addTo(map)
 
-  //главный маркер
-  const mainIcon = L.icon({
-    iconUrl: './img/main-pin.svg',
-    iconSize: [52, 52],
-    iconAnchor: [26, 52],
-  });
-  const mainPin = L.marker(
-    {lat, lng},
-    {
-      draggable: true,
-      icon: mainIcon,
-    });
-
-  mainPin
-    .addTo(map);//добавляет на карту главный пин
-
-  mainPin.on('moveend', (evt) => {//после перемещения пина передает координаты в поле адреса
-    inputAddress.value = `${evt.target.getLatLng().lat.toFixed(5)}, ${evt.target.getLatLng().lng.toFixed(5)}`;
+  .on('load', () => {
+    enablePage();
   });
 
+//главный маркер
+const mainIcon = L.icon({
+  iconUrl: './img/main-pin.svg',
+  iconSize: [52, 52],
+  iconAnchor: [26, 52],
+});
+const mainPin = L.marker(
+  {lat, lng},
+  {
+    draggable: true,
+    icon: mainIcon,
+  });
 
+mainPin
+  .addTo(map);//добавляет на карту главный пин
+
+mainPin.on('moveend', (evt) => {//после перемещения пина передает координаты в поле адреса
+  inputAddress.value = `${evt.target.getLatLng().lat.toFixed(5)}, ${evt.target.getLatLng().lng.toFixed(5)}`;
+});
+
+
+const showMap = function(loadedCardData) {
   if (loadedCardData) { //если метки загружены с сервера то отрисовать их
     const pinGroup = L.layerGroup().addTo(map);  //создание слоя layerGroup для простых маркеров
     loadedCardData.forEach((card) => {
@@ -73,33 +73,17 @@ const showMap = function(loadedCardData) {
   }
   //удаление слоя с простыми метками
   //markerGroup.clearLayers();
-
-
-  const form = document.querySelector('.ad-form');
-  const formResetButton = document.querySelector('.ad-form__reset');
-
-  const resetMap = () => {//очистка карты
-    lat = latCenter;
-    lng = lngCenter;
-    map.setView({lat, lng}, 12);
-    mainPin.setLatLng ({lat, lng});
-    inputAddress.value = `${lat}, ${lng}`;
-  };
-
-  const resetForm = (evt) => {//очистка формы
-    evt.preventDefault();
-    form.reset();
-    inputAddress.value = `${latCenter}, ${lngCenter}`;
-    defaultRoomSelector();
-    resetMap();
-  };
-
-
-
-
-  formResetButton.addEventListener('click', resetForm ); //обработка кнопки сброса формы
-  resetButton.addEventListener('click', resetMap);// обработка конопки ресет на карте
-
 };
 
-export {showMap, latCenter, lngCenter};
+const resetMap = () => {//очистка карты
+  lat = latCenter;
+  lng = lngCenter;
+  map.setView({lat, lng}, 12);
+  mainPin.setLatLng ({lat, lng});
+  inputAddress.value = `${lat}, ${lng}`;
+};
+
+
+resetButton.addEventListener('click', resetMap);// обработка конопки ресет на карте
+
+export {showMap, latCenter, lngCenter, resetMap};
